+++ packages/cubejs-backend-shared/src/index.ts
@@ -1,5 +1,5 @@
 export { getEnv, isDockerImage } from './env';
-export { QueryAlias } from './enums';
+export * from './enums';
 export * from './package';
 export * from './track';
 export * from './errors';