+++ packages/cubejs-server-core/src/core/types.ts
@@ -101,6 +101,7 @@ export type DatabaseType =
   | 'athena'
   | 'bigquery'
   | 'clickhouse'
+  | 'crate'
   | 'druid'
   | 'jdbc'
   | 'firebolt'