+++ packages/cubejs-testing/src/REQUIRED_ENV_VARS.ts
@@ -1,5 +1,5 @@
 /**
- * Environment variables that cannot be hardcoded in code, and instead must be specified via the cli.
+ * Environment variables that cannot be hardcoded, and instead must be specified via the cli.
  * Usually cloud db config & auth variables.
  */
 export const REQUIRED_ENV_VARS: {[key: string]: string[]} = {
@@ -15,6 +15,7 @@ export const REQUIRED_ENV_VARS: {[key: string]: string[]} = {
     'CUBEJS_DB_EXPORT_BUCKET',
     'CUBEJS_DB_BQ_CREDENTIALS',
   ],
+  crate: [],
   firebolt: [
     'CUBEJS_DB_USER',
     'CUBEJS_DB_PASS',