+++ rust/cubesql/cubesql/src/compile/rewrite/language.rs
@@ -363,6 +363,7 @@ macro_rules! variant_field_struct {
                 BuiltinScalarFunction::Trim => "Trim",
                 BuiltinScalarFunction::Upper => "Upper",
                 BuiltinScalarFunction::RegexpMatch => "RegexpMatch",
+                BuiltinScalarFunction::Coalesce => "Coalesce",
             }
         );
     };
@@ -395,6 +396,8 @@ macro_rules! variant_field_struct {
                 Operator::IsNotDistinctFrom => "IS_NOT_DISTINCT_FROM",
                 Operator::BitwiseAnd => "&",
                 Operator::BitwiseOr => "|",
+                Operator::BitwiseShiftRight => ">>",
+                Operator::BitwiseShiftLeft => "<<",
                 Operator::StringConcat => "||",
             }
         );